+++ SQL/updates/update-1.sql
@@ -1,0 +1,3 @@
+ALTER TABLE `settings` ADD `template` VARCHAR( 100 ) NOT NULL AFTER `base_url` ;
+
+UPDATE `slimblog`.`settings` SET `template` = 'default' WHERE `settings`.`id` =1;